Move env<->prefix methods to common
