+++ conda/plan.py
@@ -16,6 +16,7 @@ from collections import defaultdict, namedtuple
 from logging import getLogger
 from os.path import abspath, basename, dirname, exists, join
 
+from conda.common.path import preferred_env_to_prefix, preferred_env_matches_prefix
 from . import instructions as inst
 from .base.constants import DEFAULTS, LinkType
 from .base.context import context
@@ -466,38 +467,6 @@ def get_pinned_specs(prefix):
         return [i for i in f.read().strip().splitlines() if i and not i.strip().startswith('#')]
 
 
-def maybe_pad(name, pad="_"):
-    new_name = name if name.startswith(pad) else "%s%s" % (pad, name)
-    new_name = new_name if name.endswith(pad) else "%s%s" % (new_name, pad)
-    return new_name
-
-
-def preferred_env_to_prefix(preferred_env):
-    if preferred_env is None:
-        return context.root_dir
-    else:
-        return join(context.envs_dirs[0], maybe_pad(preferred_env, '_'))
-
-
-def prefix_to_env_name(prefix):
-    if prefix == context.root_dir:
-        return None
-    else:
-        env = prefix.split("/envs/")[1:]
-        return env
-
-
-def preferred_env_matches_prefix(preferred_env, prefix):
-    # type (str, str) -> bool
-    if preferred_env is None:
-        return True
-    prefix_dir = dirname(prefix)
-    if prefix_dir != join(context.root_dir, 'envs'):
-        return False
-    prefix_name = basename(prefix)
-    padded_preferred_env = maybe_pad(preferred_env)
-    return prefix_name == padded_preferred_env
-
 # Has one dist for each env
 DistForEnv = namedtuple('DistForEnv', ['env', 'dist'])
 # Has several dists for each prefix and the related r value
@@ -530,7 +499,7 @@ def install_actions(prefix, index, specs, force=False, only_names=None, always_c
 
 def not_requires_private_env(prefix, preferred_envs):
     if (context.prefix_specified is True or not context.prefix == context.root_dir or
-            all(preferred_env_matches_prefix(preferred_env, prefix) for
+            all(preferred_env_matches_prefix(preferred_env, prefix, context.root_dir) for
                 preferred_env in preferred_envs)):
         return True
     return False
@@ -552,16 +521,18 @@ def determine_dists_per_prefix(r, prefix, index, preferred_envs, dists_for_envs)
 
         def get_r(preferred_env):
             # don't make r for the prefix where we already have it created
-            if preferred_env_matches_prefix(preferred_env, prefix):
+            if preferred_env_matches_prefix(preferred_env, prefix, context.root_dir):
                 return r
             else:
-                return get_resolve_object(index.copy(), preferred_env_to_prefix(preferred_env))
+                return get_resolve_object(index.copy(), preferred_env_to_prefix(
+                    preferred_env, context.root_dir, context.envs_dirs))
 
         prefix_with_dists_no_deps_has_resolve = []
         for env in preferred_envs:
             dists = set(d.dist for d in dists_for_envs if d.env == env)
             prefix_with_dists_no_deps_has_resolve.append(
-                DistsForPrefix(prefix=preferred_env_to_prefix(env), r=get_r(env), dists=dists)
+                DistsForPrefix(prefix=preferred_env_to_prefix(
+                    env, context.root_dir, context.envs_dirs), r=get_r(env), dists=dists)
             )
     return prefix_with_dists_no_deps_has_resolve
